\name{chilean}
\docType{data}
\alias{chilean}
%%%%%%%%%%%%%%%% TITLE %%%%%%%%%%%%%%%%%%%
\title{Data: Chilean firm-level production data, 1996--2006}
%%%%%%%%%%%%%%%% DESCRIPTION %%%%%%%%%%%%%%%%%%%
\description{
  Sectoral subsample of Chilean firm-level production data covering 1996--2006.
}
%%%%%%%%%%%%%%%% USAGE %%%%%%%%%%%%%%%%%%%
\usage{data("chilean")}
%%%%%%%%%%%%%%%% FORMAT %%%%%%%%%%%%%%%%%%%
\format{
  A \link{data.frame} with 2,544 observations and 8 variables:
  \describe{
    \item{Y}{Log value added.}
    \item{sX}{Log capital.}
    \item{fX1}{Log skilled labour.}
    \item{fX2}{Log unskilled labour.}
    \item{pX}{Log materials.}
    \item{inv}{Log investment.}
    \item{idvar}{Firm identifier.}
    \item{timevar}{Year.}
  }
}
%%%%%%%%% REFERENCES %%%%%%%%%
\references{
  Instituto Nacional de Estadisticas de Chile, Encuesta Nacional Industrial Anual (ENIA).
}
%%%%%%%%% KEYWORDS %%%%%%%%%
\keyword{datasets}
