\name{show}
\alias{show,prod-method}
\alias{show}
\docType{methods}
%%%%%%%%%%%%%%%% TITLE %%%%%%%%%%%%%%%%%%%
\title{Print a table with parameter estimates}
%%%%%%%%%%%%%%%% DESCRIPTION %%%%%%%%%%%%%%%%%%%
\description{
  This method allwos the user to print a table with the parameter estimates of an \code{S4} \link{prod} object.
}
%%%%%%%%%%%%%%%% USAGE %%%%%%%%%%%%%%%%%%%
\usage{
  show(object)
}
%%%%%%%%%%%%%%%% ARGUMENTS %%%%%%%%%%%%%%%%%%%
\arguments{
  %
  \item{object}{
    object of class \link{prod}.
  }
  %
}
%%%%%%%%%%%%%%%% DETAILS %%%%%%%%%%%%%%%%%%%
\details{
  \code{show} accepts an \code{S4} \link{prod} object and prints a table with estimated parameters.
}
%%%%%%%%%%%%%%%% AUTHOR %%%%%%%%%%%%%%%%%%%
\author{Gabriele Rovigatti}
